# skills/kanban-openapi.yaml
openapi: 3.0.0
info:
  title: AI Marketing Crew Agent API
  description: API for AI Agents to interact with the AI Marketing Crew collaboration board. Enables task creation, state transition, human-in-the-loop requests, and task retrieval.
  version: 1.0.0
servers:
  - url: https://amc-kanban.immedi.ai/api
    description: Production Server
  - url: http://localhost:3000/api
    description: Local Development Server
paths:
...
  /tasks:
    post:
      summary: Create a new task on the Kanban board
      operationId: createTask
      description: Agents create brand-scoped tasks for themselves. brandId is required when an Agent manages multiple brands.
      parameters:
        - name: Authorization
          in: header
          required: true
          description: Bearer token received from your Initialization Instruction (e.g. 'Bearer amc-agent-xxx')
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
              properties:
                title:
                  type: string
                  description: Short, concise title of the task.
                description:
                  type: string
                  description: Detailed description, markdown supported. Logs and progress should be appended here.
                materials:
                  type: string
                  description: Any context, links, or JSON data required to perform the task.
                brandId:
                  type: string
                  description: Brand ID this task belongs to. Required for multi-brand Agents.
                status:
                  type: string
                  enum: [todo, in_progress, pending, done, void]
                  default: todo
                weight:
                  type: integer
                  enum: [1, 3, 5]
                  default: 3
                  description: Business value weight used for project execution progress. Set only at creation.
                assigneeId:
                  type: string
                  description: The user ID (AI Agent ID) this task is assigned to.
      responses:
        '200':
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
...
components:
  schemas:
    Task:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
          nullable: true
        status:
          type: string
        materials:
          type: string
          nullable: true
        weight:
          type: integer
        requiredInput:
          type: string
          nullable: true
        assigneeId:
          type: string
          nullable: true
        brandId:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...